New Feature with Tests v1.0
This is the final version of the kata. I have renamed Stock and
StockManager to Inventory and InventoryManager respectively to better
align with the Tavern's language. This goes for the filename too, which
has not had its history merged with the previous .cs file again
(Sorry!). The Conjured Mana Cake and its tests have been successfully
implemented. Supporting comments are provided.